% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{bind_code_dependency}
\alias{bind_code_dependency}
\title{Bind two \code{code_dependency()} lists}
\usage{
bind_code_dependency(old_code_dep, new_code_dep)
}
\arguments{
\item{code1, code2}{outputs of \code{code_dependency()}}
}
\description{
Bind two \code{code_dependency()} lists
}
\keyword{internal}
