% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\docType{class}
\name{chunk_comment}
\alias{chunk_comment}
\alias{chunk_comment$new}
\title{Code Chunk comment}
\format{
An \code{\link{R6Class}} generator object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
\section{Fields}{

\describe{
\item{\code{comment}}{(\code{character}) comment_to_show}
}}

\section{Methods}{

\describe{
\item{\code{initialize(comment = NULL)}}{
Create the object by setting the field \code{comment}
}
\item{\code{get_rcode(envir = parent.frame()}}{
For the code chunk with the string of the R-Code is returned.
}
\item{\code{get_eval_info()}}{
Returns a list of information about evaluation.
}
}
}

\examples{
x <- chunk_comment$new(comment = "I want a comment in this line")

x$eval()
x$get_rcode()
safe_comment <- x$get_rcode()

x <- chunk_comment$new(comment = "# I want a comment in this line")
x$eval()
x$get_rcode()
stopifnot(x$get_rcode() == safe_comment)
}
\keyword{data}
