% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_deep_clone}
\alias{chunks_deep_clone}
\title{Deep clones a chunk object}
\usage{
chunks_deep_clone(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
a deep copy of \code{chunks}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
use this function if you need to copy a \code{chunks} object as this
function makes sure all associated \code{environments} a deep copied (i.e. are independent
of the original object)
}
\examples{
x_chunk <- chunks$new()
chunks_push(expression(y <- 1), chunks = x_chunk)

# A copy of x_chunk which does not share the same environment
x_chunk_copy <- chunks_deep_clone(x_chunk)

# Add expression only into x_chunk
chunks_push(expression(y <- 2 * y), chunks = x_chunk)

# Get R code from both chunks, note it is different
chunks_get_rcode(x_chunk)
chunks_get_rcode(x_chunk_copy)

# Evaluate both chunks, not result is different
chunks_safe_eval(x_chunk)
chunks_safe_eval(x_chunk_copy)
}
