% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_eval}
\alias{chunks_eval}
\title{Evaluates all remaining chunks}
\usage{
chunks_eval(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead

You can evaluate all remaining chunks of chunks being setup in the shiny session (see \code{\link[=get_chunks_object]{get_chunks_object()}}).
The value of the last chunk being evaluated will be returned.

This method should not be evaluated multiple times, else it throws a warning. This is due to the
fact that multiple evaluations of the following second code snippet is dangerous:
\itemize{
\item{1 }{a <- 2}
\item{2 }{a <- a + 1}
}
\code{eval_chunks} evaluates code just once and warn user if one tries to evaluate multiple times.
}
