% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_get_eval_msg}
\alias{chunks_get_eval_msg}
\title{Get chunks evaluation message}
\usage{
chunks_get_eval_msg(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
(\code{character}) chunks evaluation message
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
