% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_get_rcode}
\alias{chunks_get_rcode}
\title{Returns the R-Code from a \code{chunks} object}
\usage{
chunks_get_rcode(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
The R code stored inside the chunks that can be used to reproduce the value of the evaluation
of all code chunks.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead

This function returns a list of the R-Code that reproduces all currently registered chunks
inside the chunk stack.
}
