% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_new}
\alias{chunks_new}
\title{Creates and returns a R6 chunks object to be used in a shiny/teal app}
\usage{
chunks_new(envir = new.env())
}
\arguments{
\item{envir}{(\code{environment} or \code{NULL}) optional, environment to get objects from to chunks environment}
}
\value{
R6 chunks object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
\examples{
new_chunks <- chunks_new()
}
\references{
chunks
}
