% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_push_data_merge}
\alias{chunks_push_data_merge}
\title{Pushes a merged_dataset to chunks}
\usage{
chunks_push_data_merge(x, chunks = get_chunks_object())
}
\arguments{
\item{x}{(\code{list}) outcome of \code{data_merge_srv}}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
