% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_reset}
\alias{chunks_reset}
\title{Resets the current chunk list}
\usage{
chunks_reset(envir = parent.frame(), chunks = get_chunks_object())
}
\arguments{
\item{envir}{(\code{environment}) The environment to hand over to the chunks.}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
nothing, it modifies shiny session object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead

Empty the chunk list and the remaining and evaluated list.
}
