% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_safe_eval}
\alias{chunks_safe_eval}
\title{Evaluates all remaining chunks and validate if error}
\usage{
chunks_safe_eval(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

You can evaluate all remaining chunks of chunks being setup in  the shiny session (see \code{\link[=get_chunks_object]{get_chunks_object()}}).
The value of the last chunk being evaluated will be returned.
}
