% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_uneval}
\alias{chunks_uneval}
\title{Resets the evaluation of the chunks.}
\usage{
chunks_uneval(
  chunks = get_chunks_object(),
  overwrite = FALSE,
  envir = parent.frame()
)
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}

\item{overwrite}{(\code{logical}) Whether to use values from the current environment
or just use a new one}

\item{envir}{(\code{environment}) An environment to use overwriting. Only used if
\code{overwrite} is \code{TRUE}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Keep code pieces.
}
