% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_validate_is}
\alias{chunks_validate_is}
\title{Raise shiny validate error if variable is not of a certain class}
\usage{
chunks_validate_is(var, class, msg = NULL, chunks = get_chunks_object())
}
\arguments{
\item{var}{(\code{character}) variable name}

\item{class}{(\code{character}) class name}

\item{msg}{optional, (\code{character}) custom error message, if \code{NULL} then default error message is used}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
shiny validation error if variable is not of certain class
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
