% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_validate_is_ok}
\alias{chunks_validate_is_ok}
\title{Raise shiny validate error if chunks status is not ok}
\usage{
chunks_validate_is_ok(msg = NULL, chunks = get_chunks_object())
}
\arguments{
\item{msg}{optional, (\code{character}) custom error message, if \code{NULL} then default error message is used}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
shiny validation error if chunks are not ok
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
\references{
chunks_is_ok
}
