% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{detect_symbol}
\alias{detect_symbol}
\title{Detects \code{"SYMBOL"} tokens for rbinded \code{getParseData()} structure}
\usage{
detect_symbol(object, pd = calls_pd)
}
\arguments{
\item{object}{\code{character} containing the name of the object}

\item{pd}{\code{list} of data.frames of results of \code{utils::getParseData()} trimmed to unique \code{parsed_code} calls}
}
\description{
Detects \code{"SYMBOL"} tokens for rbinded \code{getParseData()} structure
}
\keyword{internal}
