% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{.copy_env}
\alias{.copy_env}
\title{Copy environment content to the other environment
To avoid sharing the same environment by multiple \code{Quosure} objects}
\usage{
.copy_env(env)
}
\description{
Copy environment content to the other environment
To avoid sharing the same environment by multiple \code{Quosure} objects
}
\keyword{internal}
