% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-utils.R
\name{.keep_code_name_unique}
\alias{.keep_code_name_unique}
\title{Combine two code vectors and keep unique names}
\usage{
.keep_code_name_unique(x, y = character(0))
}
\arguments{
\item{x}{(\code{character}) vector of code.}

\item{y}{(\code{character}) vector of code.}
}
\value{
\code{character} containing both \code{x} and \code{y} combined with
unique names.
}
\description{
Combines two code vectors and modifies duplicated names.
}
\keyword{internal}
