% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{.resolve_code_vector}
\alias{.resolve_code_vector}
\title{Needed to determine if the \code{Quosure}'s code can be joined}
\usage{
.resolve_code_vector(code, code2)
}
\description{
Needed to determine if the \code{Quosure}'s code can be joined
}
\keyword{internal}
