% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{eval_code}
\alias{eval_code}
\alias{eval_code,Quosure,character-method}
\alias{eval_code,Quosure,language-method}
\alias{eval_code,Quosure,expression-method}
\title{Evaluate the code in the \code{Quosure} environment}
\usage{
eval_code(object, code, name = "code")

\S4method{eval_code}{Quosure,character}(object, code, name = "code")

\S4method{eval_code}{Quosure,language}(object, code, name = "code")

\S4method{eval_code}{Quosure,expression}(object, code, name = "code")
}
\arguments{
\item{object}{(\code{Quosure})}

\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Also acceptps and stores comments}

\item{name}{(\code{character(1)}) name of the code element.}
}
\description{
Given code is evaluated in the \code{Quosure} environment and appended to the \code{code} slot. This means
that state of the environment is always a result of the stored code (if \code{Quosure} was initialized)
with reproducible code.
}
