% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{get_chunks_object}
\alias{get_chunks_object}
\title{Gets chunks object}
\usage{
get_chunks_object(session = get_session_object(), parent_idx = 0L)
}
\arguments{
\item{session}{optional, (\code{ShinySession})\cr
\code{shiny} session object, if missing then
\link[shiny]{getDefaultReactiveDomain} is used.}

\item{parent_idx}{optional (\code{integer})\cr
from which shiny parent environment chunks object should be taken.}
}
\value{
\code{chunks} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
