% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{get_code,Quosure-method}
\alias{get_code,Quosure-method}
\title{Get the code from the \code{Quosure}}
\usage{
\S4method{get_code}{Quosure}(object)
}
\arguments{
\item{object}{(\code{Quosure})}
}
\description{
Get the code from the \code{Quosure}
}
