% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{get_code_dependencies}
\alias{get_code_dependencies}
\title{Return the lines of code (with side-effects) needed to reproduce the object.}
\usage{
get_code_dependencies(qenv, name)
}
\arguments{
\item{qenv}{\code{qenv} object}

\item{name}{\code{character} with object name}
}
\value{
\code{character} vector of elements of \code{parsed_code} calls that were required to build the side-effects and
influencing objects having and impact on the \code{object}
}
\description{
Return the lines of code (with side-effects) needed to reproduce the object.
}
\keyword{internal}
