% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_details.R
\name{get_eval_details_srv}
\alias{get_eval_details_srv}
\title{Server part of \code{get_chunks_info}}
\usage{
get_eval_details_srv(id, chunks)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{chunks}{(\code{chunks})\cr
object, from which the messages, warnings and errors are gathered}
}
\value{
an observer for \code{evaluation_details actionButton}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
