% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_details.R
\name{get_eval_details_ui}
\alias{get_eval_details_ui}
\title{The \code{UI} part of get chunks info module}
\usage{
get_eval_details_ui(id)
}
\arguments{
\item{id}{(\code{character})\cr
id of a shiny module}
}
\value{
(\code{shiny.tag})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
