% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_warnings.R
\name{get_messages}
\alias{get_messages}
\title{Get the messages of qenv object}
\usage{
get_messages(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
character vector of messages
}
\description{
Get the messages of qenv object
}
\examples{
data_q <- new_qenv() \%>\% eval_code(bquote(message("This is a message")))
get_messages(data_q)

}
