% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_details.R
\name{get_number_in_box}
\alias{get_number_in_box}
\title{Creates a \code{shiny::tag} element, which represents a colored box
with a character inside of it.}
\usage{
get_number_in_box(character, box_color, char_color)
}
\arguments{
\item{character}{(\code{character} or \code{numeric} value)\cr
a single string put inside a box}

\item{box_color}{(\code{character} value)\cr
hex code of a color (including #) or a recognized CSS3 color name}

\item{char_color}{(\code{character} value)\cr
hex code of a color (including #) or a recognized CSS3 color name}
}
\value{
\code{shiny::tag} with a character of color \code{char_color} inside
a box of color \code{box_color}
}
\description{
Creates a \code{shiny::tag} element, which represents a colored box
with a character inside of it.
}
\details{
Uses the Font-Awesome library of icons to get the box icon.
}
\keyword{internal}
