% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-get_outputs.R
\name{get_outputs}
\alias{get_outputs}
\title{Get outputs}
\usage{
get_outputs(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
list of outputs generated in a qenv
}
\description{
\code{eval_code} evaluates code silently so plots and prints don't show up in the console or graphic devices.
If one wants to use an output outside of the qenv (e.g. use a graph in \code{renderPlot}) then use \code{get_outputs}.
}
\examples{
q <- eval_code(
  qenv(),
  quote({
    a <- 1
    print("I'm an output")
    plot(1)
  })
)
get_outputs(q)
}
