% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephant.R
\name{get_var,Quosure,character-method}
\alias{get_var,Quosure,character-method}
\title{Get object from the \code{Quosure} environment}
\usage{
\S4method{get_var}{Quosure,character}(object, var)
}
\arguments{
\item{object}{(\code{Quosure})}

\item{var}{(\code{character(1)})}
}
\description{
Get object from the \code{Quosure} environment.
}
