% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-get_var.R
\name{get_var}
\alias{get_var}
\alias{get_var,Quosure,character-method}
\alias{get_var,quosure.error,ANY-method}
\alias{[[,Quosure,ANY,missing-method}
\alias{[[.quosure.error}
\title{Get object from the \code{Quosure} environment}
\usage{
get_var(object, var)

\S4method{get_var}{Quosure,character}(object, var)

\S4method{get_var}{quosure.error,ANY}(object, var)

\S4method{[[}{Quosure,ANY,missing}(x, i, j, ...)

\method{[[}{quosure.error}(x, i, j, ...)
}
\arguments{
\item{object}{(\code{Quosure})}

\item{var}{(\code{character(1)}) name of the variable to pull from the environment.}

\item{x}{(\code{Quosure})}

\item{i}{(\code{character}) name of the binding in environment (name of the variable)}

\item{j}{not used}

\item{...}{not used}
}
\description{
Get object from the \code{Quosure} environment.
}
\examples{
q1 <- new_quosure(env = list2env(list(a = 1)), code = quote(a <- 1))
q2 <- eval_code(q1, code = "b <- a")
get_var(q2, "b")
q2[["b"]]

}
