% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_warn_message_util}
\alias{get_warn_message_util}
\title{Obtain warnings or messages from code slot}
\usage{
get_warn_message_util(object, what)
}
\arguments{
\item{object}{(\code{qenv})}

\item{what}{(``}
}
\description{
Obtain warnings or messages from code slot
}
