% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_warnings.R
\name{get_warnings}
\alias{get_warnings}
\title{Get the warnings of qenv object}
\usage{
get_warnings(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
character vector of warnings
}
\description{
Get the warnings of qenv object
}
\examples{
data_q <- new_qenv() \%>\% eval_code("iris_data <- iris")
warning_qenv <- eval_code(
  data_q,
  bquote(p <- hist(iris_data[, .("Sepal.Length")], ff = ""))
)
get_warnings(warning_qenv)

}
