% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephant.R
\name{join,Quosure,Quosure-method}
\alias{join,Quosure,Quosure-method}
\title{Join two \code{Quosure} objects}
\usage{
\S4method{join}{Quosure,Quosure}(object, object2, overwrite = FALSE)
}
\arguments{
\item{object}{(\code{Quosure})}

\item{object2}{(\code{Quosure})}
}
\description{
\subsection{Combine two \code{Quosure} object by merging their environments and the code.
Not every \code{Quosure} objects can be combined:}{
}
}
