% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{join}
\alias{join}
\title{Join two \code{Quosure} objects}
\usage{
join(x, y)
}
\arguments{
\item{x}{(\code{Quosure})}

\item{y}{(\code{Quosure})}
}
\description{
Combine two \code{Quosure} object by merging their environments and the code.
Not every \code{Quosure} objects can be combined:
\itemize{
\item if their environments contains objects of the same name but not identical
\item if \code{y} has unique code element placed before common element. This means that \code{y}
in the environment of the \code{y} was evaluated some extra code before which can influence
reproducibility
\item more cases todo
}
}
