% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lang2calls}
\alias{lang2calls}
\title{Convert Language Objects to List of Calls}
\usage{
lang2calls(x)
}
\arguments{
\item{x}{\code{language} object or a list of thereof}
}
\value{
a \code{list} of \code{call}s or \code{expression}s,
Symbols and atomic vectors (which may get mixed up in a list) are returned wrapped in list.
}
\description{
Function take \code{language} objects (such as \code{expressions} or \code{calls})
and convert them into a list of \code{calls}. It's particularly useful for handling complex
\code{language} structures.
}
\keyword{internal}
