% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-names.R
\name{names.qenv}
\alias{names.qenv}
\alias{names.qenv.error}
\alias{names<-.qenv}
\alias{names<-.qenv.error}
\title{The Names of a \code{qenv} or \code{qenv_error} Object}
\usage{
\method{names}{qenv}(x, all.names = FALSE)

\method{names}{qenv.error}(x, all.names = FALSE)

\method{names}{qenv}(x) <- value

\method{names}{qenv.error}(x) <- value
}
\arguments{
\item{x}{A (\code{qenv} or \code{qenv_error}) object.}

\item{all.names}{(\code{logical(1)}) that specifies whether to include hidden
objects.}

\item{value}{Does nothing as the names assignment is not supported.}
}
\value{
A character vector of names.
}
\description{
Functions to get the names of a \code{qenv} or \code{qenv_error} object.
The names are extrapolated from the objects in the \code{qenv} environment and
are not stored statically, unlike the normal behavior of \code{names()} function.
}
\details{
Objects named with a \code{.} (dot) prefix will be ignored and not returned,
unless \code{all.names} parameter is set to \code{TRUE}.
}
\seealso{
\code{\link[base:names]{base::names()}}
}
