% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesttemplate.R
\docType{package}
\name{nesttemplate-package}
\alias{nesttemplate}
\alias{nesttemplate-package}
\title{nesttemplate: `pkgdown` template for NEST packages}
\description{
\code{nesttemplate} provides a custom \href{https://pkgdown.r-lib.org}{\code{pkgdown}}
template for NEST packages. Please don’t use it for your own package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/nesttemplate}
  \item \url{https://insightsengineering.github.io/nesttemplate/}
  \item Report bugs at \url{https://github.com/insightsengineering/nesttemplate/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

}
\keyword{internal}
