% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-constructor.R
\name{new_qenv}
\alias{new_qenv}
\alias{new_qenv,environment,expression-method}
\alias{new_qenv,environment,character-method}
\alias{new_qenv,environment,language-method}
\alias{new_qenv,environment,missing-method}
\alias{new_qenv,missing,missing-method}
\title{Deprecated function}
\usage{
new_qenv(env = new.env(parent = parent.env(.GlobalEnv)), code = character())
}
\arguments{
\item{env}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} (\code{environment})
Environment being a result of the \code{code} evaluation.}

\item{code}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
(\code{character(1)} or \code{language}) code to evaluate. Accepts and stores comments also.}
}
\value{
Errors.
}
\description{
The \code{new_qenv()} function is deprecated use \code{\link[=qenv]{qenv()}}.
}
\seealso{
\code{\link[=qenv]{qenv()}}
}
