% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{new_quosure}
\alias{new_quosure}
\alias{new_quosure,missing,missing-method}
\alias{new_quosure,character,ANY-method}
\title{Initialize \code{Quosure} object}
\usage{
new_quosure(code = character(0), env = parent.env(.GlobalEnv))

\S4method{new_quosure}{missing,missing}(code = character(0), env = parent.env(.GlobalEnv))

\S4method{new_quosure}{character,ANY}(code = character(0), env = parent.env(.GlobalEnv))
}
\arguments{
\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Also acceptps and stores comments}

\item{env}{(\code{environment}) name of the code element.}
}
\description{
Initialize \code{Quosure} object with \code{code} and \code{env}. In order to have \code{Quosure} reproducible
one needs to initialize with \code{env} which can be reproduced by the \code{code}. Alternatively, one
can create an empty \code{Quosure} and evaluate the expressions in this object.
}
