% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-class.R
\docType{class}
\name{qenv-class}
\alias{qenv-class}
\title{Reproducible class with environment and code.}
\description{
Reproducible class with environment and code.
}
\section{Slots}{

\describe{
\item{\code{code}}{(\code{expression}) to reproduce the environment}

\item{\code{code_dependency}}{A \code{list} equal to the length of \code{id}, where each element is a list containing 3 elements
\itemize{
\item \code{occurrence} - named \code{list} by object names with numeric vector as elements indicating calls in which object
appears.
\item \code{cooccurrence} - \code{list} of the same length as number of calls in \code{parsed_code}, containing \code{NULL}s if there is no
co-occurrence between objects, or a \code{character} vector indicating co-occurrence of objects in specific \code{parsed_code}
call element. If a character vector, then the first element is the name of the dependent object, and the rest are the
influencing objects
\item \code{effects} - named \code{list}  by object names with numeric vector as elements indicating which calls has effect on this
object, or NULL if there are no side-effects pointing at this object.
}}

\item{\code{env}}{(\code{environment}) environment which content was generated by the evaluation
of the \code{code} slot.}

\item{\code{id}}{(\code{integer}) random identifier of the code element to make sure uniqueness
when joining.}

\item{\code{warnings}}{(\code{character}) the warnings output when evaluating the code}

\item{\code{messages}}{(\code{character}) the messages output when evaluating the code}
}}

\keyword{internal}
