% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\docType{class}
\name{Quosure-class}
\alias{Quosure-class}
\title{Reproducible class with environment and code.}
\description{
Reproducible class with environment and code.
}
\section{Slots}{

\describe{
\item{\code{code}}{(\code{character}) to reproduce the environment}

\item{\code{env}}{(\code{environment}) environment which content was generated by the evaluation
of the \code{code} slot.}
}}

