% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_enclosing_curly_braces}
\alias{remove_enclosing_curly_braces}
\title{Removes leading and trailing curly brackets from expressions
when converting a vector of expressions to a string}
\usage{
remove_enclosing_curly_braces(code)
}
\arguments{
\item{code}{(\code{expression})\cr}
}
\value{
Character string without curly braces
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
