% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-replace_code.R
\name{replace_code}
\alias{replace_code}
\alias{replace_code,qenv,character-method}
\alias{replace_code,qenv,language-method}
\alias{replace_code,qenv,expression-method}
\alias{replace_code,qenv.error,ANY-method}
\title{Replace code}
\usage{
replace_code(object, code)
}
\arguments{
\item{object}{\code{qenv}}

\item{code}{(\code{character} or \code{language}) code to set}
}
\value{
A modified \code{qenv}.
}
\description{
Replace last element of code stored in \code{qenv}.
}
\details{
Replace the last element of the character vector stored in \verb{@code} with provided \code{code}.
This can be used to mask sensitive information, e.g. credentials.
}
\section{Examples}{

\code{q <- new_qenv()} \cr
\code{q <- eval_code(q, expression(i <- iris, ii <- subset(i, Species == "virginica")))} \cr
\code{get_code(q)} \cr
\code{q <- replace_code(q, expression(i <- iris, ii <- subset(i, Species != "virginica")))} \cr
\code{get_code(q)} \cr
}

\seealso{
\code{\linkS4class{qenv}}, \code{\link{eval_code}}
}
\keyword{internal}
