% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{return_code}
\alias{return_code}
\title{Return the lines of code needed to reproduce the object.}
\usage{
return_code(
  object,
  pd = calls_pd,
  occur = occurence,
  cooccur = cooccurence,
  parent = NULL
)
}
\arguments{
\item{object}{\code{character} with object name}

\item{occur}{result of \code{code_dependency()$occurence}}

\item{cooccur}{result of \code{code_dependency()$cooccurence}}

\item{parent}{\code{NULL} or \code{numeric} vector - in a recursive call, it is possible needed to drop parent object indicator to
omit dependency cycles}
}
\value{
\code{numeric} vector indicating which lines of \code{parsed_code} calls are required to build the \code{object}
}
\description{
Return the lines of code needed to reproduce the object.
}
\keyword{internal}
