% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{return_code_for_effects}
\alias{return_code_for_effects}
\title{Return the lines of code needed to reproduce the side-effects having an impact on the object.}
\usage{
return_code_for_effects(
  object,
  pd = calls_pd,
  occur = occurrence,
  cooccur = cooccurrence
)
}
\arguments{
\item{object}{\code{character} with object name}

\item{pd}{\code{list} of data.frames of results of \code{utils::getParseData()} trimmed to unique \code{parsed_code} calls}

\item{occur}{result of \code{code_dependency()$occurrence}}

\item{cooccur}{result of \code{code_dependency()$cooccurrence}}
}
\value{
\code{numeric} vector indicating which lines of \code{parsed_code} calls are required to build the side-effects having
and impact on the \code{object}
}
\description{
Return the lines of code needed to reproduce the side-effects having an impact on the object.
}
\keyword{internal}
