% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-show.R
\name{show,Quosure-method}
\alias{show,Quosure-method}
\title{Show the \code{Quosure} object}
\usage{
\S4method{show}{Quosure}(object)
}
\arguments{
\item{object}{(\code{Quosure})}
}
\value{
nothing
}
\description{
Prints the \code{Quosure} object
}
\examples{
q1 <- new_quosure(code = "print('a')", env = new.env())
q1
}
