% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-extract.R
\name{subset-qenv}
\alias{subset-qenv}
\alias{[.qenv}
\title{Subsets \code{qenv}}
\usage{
\method{[}{qenv}(x, names, ...)
}
\arguments{
\item{x}{(\code{qenv})}

\item{names}{(\code{character}) names of objects included in \code{\link{qenv}} to subset. Names not present in \code{\link{qenv}}
are skipped.}

\item{...}{internal usage, please ignore.}
}
\description{
Subsets \code{\link{qenv}} environment and limits the code to the necessary needed to build limited objects.
}
\examples{
q <- qenv()
q <- eval_code(q, "a <- 1;b<-2")
q["a"]
q[c("a", "b")]

}
