% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.code-package.R
\docType{package}
\name{teal.code-package}
\alias{teal.code}
\alias{teal.code-package}
\title{\code{teal.code}: Code storage and execution class for teal applications}
\description{
The \code{teal.code} package contains functionalities that can be used
to store and execute R calls. \code{qenv} class allows to create a \code{teal}
module which returns code to reproduce application outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/teal.code/}
  \item \url{https://github.com/pharmaverse/teal.code}
  \item Report bugs at \url{https://github.com/pharmaverse/teal.code/issues}
}

}
\author{
\strong{Maintainer}: Dony Unardi \email{unardid@gene.com} [reviewer]

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com}
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item Chendi Liao \email{chendi.liao@roche.com} [reviewer]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
