% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{used_in_function}
\alias{used_in_function}
\title{Whether an object is used inside a function within a call}
\usage{
used_in_function(call, object)
}
\arguments{
\item{call}{An element of \code{calls_pd} list used in \code{detect_symbol}.}

\item{object}{A character with object name.}
}
\value{
A \code{logical(1)}.
}
\description{
Whether an object is used inside a function within a call
}
\keyword{internal}
