% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCode.R
\name{check_joinable}
\alias{check_joinable}
\title{If two \code{QCode} can be joined}
\usage{
check_joinable(x, y)
}
\value{
logical(1)
}
\description{
Checks if two \code{QCode} objects can be combined.
They can't be combined if (and):
\itemize{
\item both share the same code (identified by \code{id})
\item indices of the shared code are not consecutive or don't start from 1
}
}
