% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_is_ok}
\alias{chunks_is_ok}
\title{Check chunks status (i.e. evaluated and no errors)}
\usage{
chunks_is_ok(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
(\code{logical}) flag whether everything is good
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
