% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_new}
\alias{chunks_new}
\title{Creates and returns a R6 chunks object to be used in a shiny/teal app}
\usage{
chunks_new(envir = new.env())
}
\arguments{
\item{envir}{(\code{environment} or \code{NULL}) optional, environment to get objects from to chunks environment}
}
\value{
R6 chunks object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
new_chunks <- chunks_new()
}
\references{
chunks
}
