% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_push}
\alias{chunks_push}
\title{Pushes a code chunk for global chunks}
\usage{
chunks_push(expression, id = NULL, chunks = get_chunks_object())
}
\arguments{
\item{expression}{(\code{call}) Expression that contains a function call.}

\item{id}{optional, (\code{character}) ID given for the code chunk}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
Nothing, just add the chunk to the \code{chunks} argument
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
all_chunks <- chunks_new()
chunks_push(chunks = all_chunks, expression = call("as.character", x = 3), id = "tbl")

x <- 3
chunks_push(chunks = all_chunks, expression = quote(as.character(x)), id = "tbl2")
}
