% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_push_chunks}
\alias{chunks_push_chunks}
\title{Pushes a chunks stack to global chunks}
\usage{
chunks_push_chunks(x, overwrite = FALSE, chunks = get_chunks_object())
}
\arguments{
\item{x}{(\code{chunks}) stack object}

\item{overwrite}{optional, (\code{logical}) Whether to ignore conflicts
between the two chunk environments and use the added one as default}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\value{
Nothing, just add the chunk to the \code{chunks} argument
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
chunks_object <- chunks$new()
chunks_push(chunks = chunks_object, expression = bquote(x <- 1))

chunks_object2 <- chunks$new()
chunks_push(chunks = chunks_object2, expression = bquote(y <- 1))

chunks_push_chunks(chunks = chunks_object, x = chunks_object2)

chunks_get_rcode(chunks_object)
}
