% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_push_comment}
\alias{chunks_push_comment}
\title{Pushes a code comment chunk for global chunks}
\usage{
chunks_push_comment(comment, chunks = get_chunks_object())
}
\arguments{
\item{comment}{(code{character}) Comment to be inserted into the Show-R code.}

\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
all_chunks <- chunks_new()
chunks_push_comment("this is a comment", chunks = all_chunks)

chunks_get_rcode(chunks = all_chunks) == "# this is a comment"
}
