% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_push_new_line}
\alias{chunks_push_new_line}
\title{Adds an empty line to the code chunks}
\usage{
chunks_push_new_line(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
\examples{
all_chunks <- chunks_new()
chunks_push_new_line(chunks = all_chunks)

chunks_get_rcode(chunks = all_chunks) == " "
}
