% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_safe_eval}
\alias{chunks_safe_eval}
\title{Evaluates all remaining chunks and validate if error}
\usage{
chunks_safe_eval(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead

You can evaluate all remaining chunks of chunks being setup in the shiny session (see \code{\link[=get_chunks_object]{get_chunks_object()}}).
The value of the last chunk being evaluated will be returned.
}
