% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks_warnings}
\alias{chunks_warnings}
\title{Gets warnings from the \code{chunks} object}
\usage{
chunks_warnings(chunks = get_chunks_object())
}
\arguments{
\item{chunks}{optional, (\code{chunks}) object.
If not provided then automatic \code{chunks} object detection is run via \link{get_chunks_object}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function returns a list of all the warnings encountered during evaluation of the code
in the \code{chunks} object.
}
