% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{detect_symbol}
\alias{detect_symbol}
\title{Detects \code{"SYMBOL"} tokens for row \code{bounded} \code{getParseData()} structure}
\usage{
detect_symbol(object, pd)
}
\arguments{
\item{object}{\code{character} containing the name of the object}

\item{pd}{\code{list} of \code{data.frame}s of results of \code{utils::getParseData()} trimmed to unique \code{parsed_code} calls}
}
\value{
A \code{logical} vector pointing in which elements of \code{pd} the \code{SYMBOL} token row has \code{object} in text column
}
\description{
Detects \code{"SYMBOL"} tokens for row \code{bounded} \code{getParseData()} structure
}
\keyword{internal}
