% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-join.R
\name{.check_joinable}
\alias{.check_joinable}
\title{If two \code{Quosure} can be joined}
\usage{
.check_joinable(x, y)
}
\arguments{
\item{x}{(\code{Quosure})}

\item{y}{(\code{Quosure})}
}
\value{
\code{TRUE} if able to join or \code{character} used to print error message.
}
\description{
Checks if two \code{Quosure} objects can be combined.
They can't be combined if (and):
\itemize{
\item both share the same code (identified by \code{id})
\item indices of the shared code are not consecutive or don't start from 1
}
}
\keyword{internal}
