% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-join.R
\name{.check_joinable}
\alias{.check_joinable}
\title{If two \code{qenv} can be joined}
\usage{
.check_joinable(x, y)
}
\arguments{
\item{x}{(\code{qenv})}

\item{y}{(\code{qenv})}
}
\value{
\code{TRUE} if able to join or \code{character} used to print error message.
}
\description{
Checks if two \code{qenv} objects can be combined.
For more information, please see \code{\link{join}}
}
\keyword{internal}
