% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code-dependency.R
\name{get_children}
\alias{get_children}
\title{Get children calls within \code{getParseData()} object}
\usage{
get_children(pd, parent)
}
\arguments{
\item{pd}{\code{list} of data.frames of results of \code{utils::getParseData()} trimmed to unique \code{parsed_code} calls}

\item{parent}{parent id in \code{utils::getParseData()}}
}
\value{
Row \code{binded} \code{utils::getParseData()} of all calls.
}
\description{
Get children calls within \code{getParseData()} object
}
\keyword{internal}
