% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{get_children}
\alias{get_children}
\title{Get child calls within \code{getParseData()} object}
\usage{
get_children(pd, parent)
}
\arguments{
\item{pd}{(\code{data.frame}) A result of \code{utils::getParseData()}.}

\item{parent}{Object parent id in \code{utils::getParseData()}.}
}
\value{
Row \code{bounded} \code{utils::getParseData()} of all elements of a call pointing to a \code{parent} id.
}
\description{
Get child calls within \code{getParseData()} object
}
\keyword{internal}
