% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-get_code.R
\name{get_code}
\alias{get_code}
\alias{get_code,Quosure-method}
\alias{get_code,quosure.error-method}
\title{Get code from \code{Quosure}}
\usage{
get_code(object)

\S4method{get_code}{Quosure}(object)

\S4method{get_code}{quosure.error}(object)
}
\arguments{
\item{object}{(\code{Quosure})}
}
\value{
named \code{character} with the reproducible code.
}
\description{
Get code from \code{Quosure}
}
\examples{
q1 <- new_quosure(env = list2env(list(a = 1)), code = quote(a <- 1))
q2 <- eval_code(q1, code = "b <- a", name = "copy a to b")
q3 <- eval_code(q2, code = "# comment line")
q4 <- eval_code(q3, code = "d <- 2", name = "assign d")
get_code(q4)

}
