% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{get_code_dependency}
\alias{get_code_dependency}
\title{Return the lines of code (with side-effects) needed to reproduce the object}
\usage{
get_code_dependency(code, names)
}
\arguments{
\item{code}{An \code{expression} with \code{srcref} attribute or a \code{character} with the code.}

\item{names}{A \code{character(n)} with object names.}
}
\value{
\code{character} vector of elements of \code{code} calls that were required to build the side-effects and
influencing objects having and impact on the \code{object}
}
\description{
Return the lines of code (with side-effects) needed to reproduce the object
}
\keyword{internal}
