% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-get_env.R
\name{get_env}
\alias{get_env}
\alias{get_env,qenv-method}
\alias{get_env,qenv.error-method}
\title{Access environment included in \code{qenv}}
\usage{
get_env(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
An \code{environment} stored in \code{qenv@env} slot.
}
\description{
The access of environment included in \code{qenv@env} allows to e.g. list object names included in \code{qenv@env} slot.
}
\examples{
q <- qenv()
q1 <- within(q, {
  a <- 5
  b <- data.frame(x = 1:10)
})
get_env(q1)
ls(get_env(q1))

}
