% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-get_var.R
\name{get_var}
\alias{get_var}
\alias{get_var,qenv,character-method}
\alias{get_var,qenv.error,ANY-method}
\title{Get object from \code{qenv}}
\usage{
get_var(object, var)
}
\arguments{
\item{object, x}{(\code{qenv})}

\item{var, i}{(\code{character(1)}) variable name.}
}
\value{
The value of required variable (\code{var}) within \code{qenv} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} by native \R operators/functions:
\code{x[[name]]}, \code{x$name} or \code{\link[=get]{get()}}.

Retrieve variables from the \code{qenv} environment.
}
\examples{
q <- qenv()
q1 <- eval_code(q, code = quote(a <- 1))
q2 <- eval_code(q1, code = "b <- a")
get_var(q2, "b")

}
