% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-get_var.R
\name{get_var}
\alias{get_var}
\alias{[[.qenv.error}
\title{Get object from \code{qenv}}
\usage{
get_var(...)

\method{[[}{qenv.error}(x, i)
}
\arguments{
\item{...}{function is deprecated.}

\item{x}{(\code{qenv})}

\item{i}{(\code{character(1)}) variable name.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Instead of \code{\link[=get_var]{get_var()}} use native \R operators/functions:
\code{x[[name]]}, \code{x$name} or \code{\link[=get]{get()}}:
}
