% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-get_var.R
\name{get_var}
\alias{get_var}
\alias{get_var,qenv,character-method}
\alias{get_var,qenv.error,ANY-method}
\alias{[[,qenv-method}
\title{Get object from \code{qenv}}
\usage{
get_var(object, var)

\S4method{[[}{qenv}(x, i)
}
\arguments{
\item{object, x}{(\code{qenv})}

\item{var, i}{(\code{character(1)}) variable name.}
}
\value{
The value of required variable (\code{var}) within \code{qenv} object.
}
\description{
Retrieve variables from the \code{qenv} environment.
}
\examples{
q <- qenv()
q1 <- eval_code(q, code = quote(a <- 1))
q2 <- eval_code(q1, code = "b <- a")
get_var(q2, "b")
q2[["b"]]

}
