% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_warn_message_util}
\alias{get_warn_message_util}
\title{Obtain warnings or messages from code slot}
\usage{
get_warn_message_util(object, what)
}
\arguments{
\item{object}{(\code{qenv})}

\item{what}{(\code{warning} or \code{message})}
}
\value{
\code{character(1)} containing combined message or \code{NULL} when no warnings/messages
}
\description{
Obtain warnings or messages from code slot
}
\keyword{internal}
