% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{init_chunks}
\alias{init_chunks}
\title{Allows using chunks in the global environment of a teal app}
\usage{
init_chunks(new_chunks = chunks_new(), session = get_session_object())
}
\arguments{
\item{new_chunks}{optional, (\code{chunks_stack}) new object to initialize with}

\item{session}{optional, (\code{ShinySession}) \code{shiny} session object.}
}
\value{
nothing, it modifies shiny session object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
\references{
chunks
}
