% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-join.R
\name{join}
\alias{join}
\alias{join,Quosure,Quosure-method}
\alias{join,QuosureError,ANY-method}
\alias{join,ANY,QuosureError-method}
\title{Join two \code{Quosure} objects}
\usage{
join(x, y)

\S4method{join}{Quosure,Quosure}(x, y)

\S4method{join}{QuosureError,ANY}(x, y)

\S4method{join}{ANY,QuosureError}(x, y)
}
\arguments{
\item{x}{(\code{Quosure})}

\item{y}{(\code{Quosure})}
}
\description{
Combine two \code{Quosure} object by merging their environments and the code.
Not every \code{Quosure} objects can be combined:
\itemize{
\item if their environments contains objects of the same name but not identical
\item if \code{y} has unique code element placed before common element. This means that \code{y}
in the environment of the \code{y} was evaluated some extra code before which can influence
reproducibility
\item more cases to be done
}
}
\examples{
q1 <- new_quosure(
  code = c(iris1 = "iris1 <- iris", mtcars1 = "mtcars1 <- mtcars"),
  env = list2env(list(
    iris1 = iris,
    mtcars1 = mtcars
  ))
)
q2 <- q1
q1 <- eval_code(q1, "iris2 <- iris")
q2 <- eval_code(q2, "mtcars2 <- mtcars")
qq <- join(q1, q2)
get_code(qq)
}
