% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-join.R
\name{join}
\alias{join}
\alias{join,qenv,qenv-method}
\alias{join,qenv.error,ANY-method}
\alias{join,qenv,qenv.error-method}
\title{Join two \code{qenv} objects}
\usage{
join(x, y, force, verbose)

\S4method{join}{qenv,qenv}(x, y, force, verbose)

\S4method{join}{qenv.error,ANY}(x, y)

\S4method{join}{qenv,qenv.error}(x, y)
}
\arguments{
\item{x}{(\code{qenv})}

\item{y}{(\code{qenv})}

\item{force}{(\code{logical(1)}) whether to turn off automatic validation and duplicated code removal, by default \code{FALSE}.
Please be careful as the default setup is recommended and supported,
the forced join is available only for advanced users who are aware of all risks.}

\item{verbose}{(\code{logical(1)}) whether to print messages and warnings connected with the \code{force} argument.
By default to \code{TRUE}.}
}
\description{
Combine two \code{qenv} objects by merging their environments and the code.
Not all \code{qenv} objects can be combined:
\itemize{
\item if their environments contains objects of the same name but not identical
\item if \code{y} has unique code element placed before common element. This means that \code{y}
in the environment of the \code{y} was evaluated some extra code before which can influence
reproducibility
\item more cases to be done
}
}
\examples{
q1 <- new_qenv(
  code = c(iris1 = "iris1 <- iris", mtcars1 = "mtcars1 <- mtcars"),
  env = list2env(list(
    iris1 = iris,
    mtcars1 = mtcars
  ))
)
q2 <- q1
q1 <- eval_code(q1, "iris2 <- iris")
q2 <- eval_code(q2, "mtcars2 <- mtcars")
qq <- join(q1, q2)
get_code(qq)
}
