% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-c.R, R/qenv-join.R
\name{c.qenv}
\alias{c.qenv}
\alias{c.qenv.error}
\alias{join}
\title{Join \code{qenv} objects}
\usage{
\method{c}{qenv}(...)

\method{c}{qenv.error}(...)

join(...)
}
\arguments{
\item{...}{function is deprecated.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Instead of \code{\link[=join]{join()}} use \code{\link[=c]{c()}}.
}
\examples{
q <- qenv()
q1 <- within(q, {
  iris1 <- iris
  mtcars1 <- mtcars
})
q1 <- within(q1, iris2 <- iris)
q2 <- within(q1, mtcars2 <- mtcars)
qq <- c(q1, q2)
cat(get_code(qq))

}
