% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lang2calls}
\alias{lang2calls}
\title{recursively convert language object to list of simple calls}
\usage{
lang2calls(x)
}
\arguments{
\item{x}{a call or a list of calls}
}
\description{
recursively convert language object to list of simple calls
}
\keyword{internal}
