% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-names.R
\name{names.qenv}
\alias{names.qenv}
\alias{names.qenv.error}
\title{The Names of a \code{qenv} or \code{qenv_error} Object}
\usage{
\method{names}{qenv}(x, ...)

\method{names}{qenv.error}(x)
}
\arguments{
\item{x}{(\code{qenv} or \code{qenv_error}) object.}

\item{...}{Additional parameters to this function, allowed parameters:
\itemize{
\item \code{all.names}: (\code{logical(1)}) that specifies whether to include hidden objects.
}}
}
\value{
A character vector of names.
}
\description{
Functions to get the names of a \code{qenv} or \code{qenv_error} object.
The names are extrapolated from the objects in the \code{qenv} environment and
are not stored statically, unlike the normal behavior of \code{names()} function.
}
\details{
Objects named with a \code{.} (dot) prefix will be ignored and not returned,
unless \code{all.names} parameter is set to \code{TRUE}.
}
\examples{
q1 <- within(qenv(), iris <- iris)
names(q1)

q2 <- within(q1, {
  mtcars <- mtcars
  CO2 <- CO2
})
names(q2)

}
\seealso{
\code{\link[base:names]{base::names()}}
}
