% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-constructor.R
\name{new_qenv}
\alias{new_qenv}
\title{Initialize \code{qenv} object}
\usage{
new_qenv()
}
\value{
\code{qenv} object.
}
\description{
Create an empty \code{qenv} object.
}
\details{
Instantiates a \code{qenv} with an empty environment.
Any changes must be made by evaluating code in it with \code{eval_code} or \code{within},
thereby ensuring reproducibility.
}
\examples{
new_qenv()

}
