% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quosure.R
\name{new_quosure}
\alias{new_quosure}
\alias{new_quosure,character,environment-method}
\alias{new_quosure,language,environment-method}
\alias{new_quosure,expression,environment-method}
\alias{new_quosure,ANY,environment-method}
\alias{new_quosure,ANY,list-method}
\alias{new_quosure,missing,missing-method}
\title{Initialize \code{Quosure} object}
\usage{
new_quosure(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{character,environment}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{language,environment}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{expression,environment}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{ANY,environment}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{ANY,list}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)

\S4method{new_quosure}{missing,missing}(
  code = character(0),
  env = new.env(parent = parent.env(.GlobalEnv))
)
}
\arguments{
\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Accepts and stores comments also.}

\item{env}{(\code{environment}) Environment being a result of the \code{code} evaluation.}
}
\description{
Initialize \code{Quosure} object with \code{code} and \code{env}. In order to have \code{Quosure} reproducible
one needs to initialize with \code{env} which can be reproduced by the \code{code}. Alternatively, one
can create an empty \code{Quosure} and evaluate the expressions in this object.
}
