% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure-constructor.R
\name{new_quosure}
\alias{new_quosure}
\alias{new_quosure,environment,character-method}
\alias{new_quosure,environment,expression-method}
\alias{new_quosure,environment,language-method}
\alias{new_quosure,list,ANY-method}
\alias{new_quosure,missing,missing-method}
\title{Initialize \code{Quosure} object}
\usage{
new_quosure(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)

\S4method{new_quosure}{environment,character}(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)

\S4method{new_quosure}{environment,expression}(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)

\S4method{new_quosure}{environment,language}(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)

\S4method{new_quosure}{list,ANY}(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)

\S4method{new_quosure}{missing,missing}(
  env = new.env(parent = parent.env(.GlobalEnv)),
  code = expression()
)
}
\arguments{
\item{env}{(\code{environment}) Environment being a result of the \code{code} evaluation.}

\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Accepts and stores comments also.}
}
\description{
Initialize \code{Quosure} object with \code{code} and \code{env}. In order to have \code{Quosure} reproducible
one needs to initialize with \code{env} which can be reproduced by the \code{code}. Alternatively, one
can create an empty \code{Quosure} and evaluate the expressions in this object using \code{eval_code}.
}
\examples{
new_quosure(env = list2env(list(a = 1)), code = quote(a <- 1))
new_quosure(env = list2env(list(a = 1)), code = parse(text = "a <- 1"))

}
