% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{overwrite_chunks}
\alias{overwrite_chunks}
\title{Overwrites chunks object in session}
\usage{
overwrite_chunks(
  x = chunks_new(envir = parent.frame()),
  session = get_session_object()
)
}
\arguments{
\item{x}{(\code{chunks}) \link{chunks}-object to be used inside the current
session.}

\item{session}{optional, (\code{ShinySession}) \code{shiny} session object,
if missing then \link[shiny]{getDefaultReactiveDomain} is used.}
}
\value{
nothing, it modifies shiny session object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Chunks are being deprecated \code{qenv} objects should be used instead
}
