% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-class.R
\docType{class}
\name{qenv-class}
\alias{qenv-class}
\title{Reproducible class with environment and code}
\description{
Reproducible class with environment and code.
}
\section{Slots}{

\describe{
\item{\code{code}}{(\code{list} of \code{character}) representing code necessary to reproduce the environment.
Read more in Code section.}

\item{\code{env}}{(\code{environment}) environment which content was generated by the evaluation
of the \code{code} slot.}
}}

\section{Code}{


Each code element is a character representing one call. Each element has possible attributes:
\itemize{
\item warnings (\code{character}) the warnings output when evaluating the code element
\item messages (\code{character}) the messages output when evaluating the code element
\item id (\code{integer}) random identifier of the code element to make sure uniqueness when joining.
}
}

\keyword{internal}
