% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-class.R
\docType{class}
\name{qenv-class}
\alias{qenv-class}
\title{Reproducible class with environment and code}
\description{
Reproducible class with environment and code.
}
\section{Slots}{

\describe{
\item{\code{.xData}}{(\code{environment}) environment with content was generated by the evaluation}

\item{\code{code}}{(\verb{named list} of \code{character}) representing code necessary to reproduce the environment.
Read more in Code section.
of the \code{code} slot.}
}}

\section{Code}{


Each code element is a character representing one call. Each element is named with the random
identifier to make sure uniqueness when joining. Each element has possible attributes:
\itemize{
\item \code{warnings} (\code{character}) the warnings output when evaluating the code element.
\item \code{messages} (\code{character}) the messages output when evaluating the code element.
\item \code{dependency} (\code{character}) names of objects that appear in this call and gets affected by this call,
separated by \verb{<-} (objects on LHS of \verb{<-} are affected by this line, and objects on RHS are affecting this line).
}
}

\keyword{internal}
