% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{return_code}
\alias{return_code}
\title{Return the lines of code needed to reproduce the object.}
\usage{
return_code(object, occur, cooccur, eff, parent = NULL)
}
\arguments{
\item{object}{\code{character} with object name}

\item{occur}{result of \code{code_dependency()$occurrence}}

\item{cooccur}{result of \code{code_dependency()$cooccurrence}}

\item{eff}{result of \code{code_dependency()$effects}}

\item{parent}{\code{NULL} or \code{numeric} vector - in a recursive call, it is possible needed to drop parent object
indicator to omit dependency cycles}
}
\value{
\code{numeric} vector indicating which lines of \code{parsed_code} calls are required to build the \code{object}

A \code{numeric} vector with number of lines of input \code{pd} to be returned.
}
\description{
Return the lines of code needed to reproduce the object.
}
\keyword{internal}
