% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-code_dependency.R
\name{return_code_for_effects}
\alias{return_code_for_effects}
\title{Return the lines of code needed to reproduce the side-effects having an impact on the object.}
\usage{
return_code_for_effects(object, calls_pd, occur, cooccur, eff)
}
\arguments{
\item{object}{\code{character} with object name}

\item{calls_pd}{A \code{list} of \code{data.frame}s, which is a result of \code{get_children(utils::getParseData(), parent = 0)}
applied on \code{parse(text = code, keep.source = TRUE)} at \code{code_dependency(code)}.}

\item{occur}{result of \code{code_dependency()$occurrence}}

\item{cooccur}{result of \code{code_dependency()$cooccurrence}}
}
\value{
\code{numeric} vector indicating which lines of \code{parsed_code} calls are required to build the side-effects having
and impact on the \code{object}

A \code{numeric} vector with number of lines of input \code{pd} to be returned for effects.
}
\description{
Return the lines of code needed to reproduce the side-effects having an impact on the object.
}
\keyword{internal}
