% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-show.R
\name{show,qenv-method}
\alias{show,qenv-method}
\title{Show the \code{qenv} object}
\usage{
\S4method{show}{qenv}(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
nothing
}
\description{
Prints the \code{qenv} object
}
\examples{
q1 <- new_qenv(
  code = "a <- 5
    b <- data.frame(x = 1:10)",
  env = list2env(list(a = 5, b = data.frame(x = 1:10)))
)
q1
}
