% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qenv-show.R
\name{show,qenv-method}
\alias{show,qenv-method}
\alias{show-qenv}
\title{Display \code{qenv} object}
\usage{
\S4method{show}{qenv}(object)
}
\arguments{
\item{object}{(\code{qenv})}
}
\value{
\code{object}, invisibly.
}
\description{
Prints the \code{qenv} object.
}
\examples{
q <- qenv()
q1 <- eval_code(q, expression(a <- 5, b <- data.frame(x = 1:10)))
q1

}
