% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_details.R
\name{show_eval_details_modal}
\alias{show_eval_details_modal}
\title{Shows Evaluation Details Modal}
\usage{
show_eval_details_modal(chunks)
}
\arguments{
\item{chunks}{(\code{chunks})\cr
object, from which the messages, warnings and errors are gathered}
}
\value{
\code{shiny.tag}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Use the \code{\link[shiny:showModal]{shiny::showModal()}} function to show the errors generated from chunks.
}
\references{
\code{\link[shiny:showModal]{shiny::showModal()}} \code{\link[=get_eval_details_srv]{get_eval_details_srv()}}
}
