% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.code-package.R
\docType{package}
\name{teal.code-package}
\alias{teal.code}
\alias{teal.code-package}
\title{teal.code: Code storage and execution class for teal applications}
\description{
The \code{teal.code} package contains functionalities that can be used
to store and execute R calls. \code{\link{chunks}} class allows to create a \code{teal}
module which returns code to reproduce application outputs.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Pawel Rucki \email{pawel.rucki@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item Tadeusz Lewandowski \email{tadeusz.lewandowski@roche.com} [contributor]
  \item Roche [copyright holder, funder]
}

}
\keyword{internal}
