% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.code-package.R
\docType{package}
\name{teal.code-package}
\alias{teal.code}
\alias{teal.code-package}
\title{teal.code: Code storage and execution class for teal applications}
\description{
The \code{teal.code} package contains functionalities that can be used
to store and execute R calls. \code{\link{chunks}} class allows to create a \code{teal}
module which returns code to reproduce application outputs.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
